+++ openspoor/__init__.py
@@ -1,5 +1,3 @@
-from openspoor.transformers.TransformerCoordinatesToSpoor import *
-from openspoor.transformers.TransformerGeocodeToCoordinates import *
-from openspoor.transformers.TransformerSpoortakToCoordinates import *
-from openspoor.mapservices.PUICMapservices import *
-from openspoor.mapservices.SpoortakMapservices import *
+from importlib_metadata import version
+
+__version__ = version('openspoor')
