+++ openspoor/mapservices/MapservicesData.py
@@ -9,8 +9,8 @@ from loguru import logger
 from shapely.geometry import Point, LineString, Polygon
 import pickle
 
-from openspoor.utils.map_services_requests import secure_map_services_request
-from openspoor.utils.common import read_config
+from ..utils.map_services_requests import secure_map_services_request
+from ..utils.common import read_config
 
 config = read_config()
