+++ openspoor/mapservices/PUICMapservices.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from loguru import logger
-from openspoor.mapservices.MapservicesData import MapservicesData
-from openspoor.utils.common import read_config
+from .MapservicesData import MapservicesData
+from ..utils.common import read_config
 
 config = read_config()
