+++ openspoor/mapservices/__init__.py
@@ -1,0 +1,2 @@
+from .PUICMapservices import *
+from .SpoortakMapservices import *
